chore: refine previous C programs and fix logic/comments

+++ 01_Basics/16_areaOfCircle_HW.c
@@ -2,7 +2,7 @@
 
 int main()
 {
-    int r = 5; 
+    float r = 5; 
     float a = 3.1416 * r * r; 
     printf("Area of circle : %f", a); 
 }
